requirements.txt
google-genai
google-generativeai
gTTS
python-dotenv
Pillow
streamlit